Merge in KI_ATLAS/atlas from feature/ATLAS-1085-traffic-point-elements-write to master

* commit '9535c07fc342d8019e1debcd3d470a8ec274c1db':
  ATLAS-1085: Rename db migration filename.
  ATLAS-1085: Add user rights check tests.
  ATLAS-1085: Improve rights check.
  ATLAS-1085: Adjustments based on PR comments.
  ATLAS-1085: Add service point number validation check for traffic point import.
  ATLAS-1085: Fix failing test.
  ATLAS-1085: Few more changes for traffic point fields.
  ATLAS-1085: Add changes for traffic point fields.
  ATLAS-1085: Add permission check for create and update.
  ATLAS:1085: Add trafficpoint versioning tests.
  ATLAS-1085: Check if service point number exists.
  ATLAS-1085: Add update check.
  ATLAS-1085: Add tests for create and update traffic points.
  ATLAS-1085: Add create traffic point test.
  ATLAS-1085: Cleanup non used code.
  ATLAS-1085: Add changes for create TrafficPoint.
  ATLAS-1085: WIP create traffic point.
  ATLAS-1085: Create and update traffic point, initial commit.

+++ service-point-directory/src/main/java/ch/sbb/atlas/servicepointdirectory/service/trafficpoint/TrafficPointElementService.java
@@ -1,25 +1,31 @@
 package ch.sbb.atlas.servicepointdirectory.service.trafficpoint;
 
+import ch.sbb.atlas.servicepointdirectory.entity.ServicePointVersion;
 import ch.sbb.atlas.servicepointdirectory.entity.TrafficPointElementVersion;
 import ch.sbb.atlas.servicepointdirectory.model.search.TrafficPointElementSearchRestrictions;
 import ch.sbb.atlas.servicepointdirectory.repository.TrafficPointElementVersionRepository;
-import ch.sbb.atlas.servicepointdirectory.service.BasePointUtility;
 import ch.sbb.atlas.versioning.model.VersionedObject;
 import ch.sbb.atlas.versioning.service.VersionableService;
-import java.util.List;
-import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.hibernate.StaleObjectStateException;
 import org.springframework.data.domain.Page;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+import java.util.Optional;
 
 @Service
 @Slf4j
 @RequiredArgsConstructor
+@Transactional
 public class TrafficPointElementService {
 
   private final TrafficPointElementVersionRepository trafficPointElementVersionRepository;
   private final VersionableService versionableService;
+  private final TrafficPointElementValidationService trafficPointElementValidationService;
 
   public Page<TrafficPointElementVersion> findAll(TrafficPointElementSearchRestrictions searchRestrictions) {
     return trafficPointElementVersionRepository.findAll(searchRestrictions.getSpecification(), searchRestrictions.getPageable());
@@ -37,7 +43,14 @@ public class TrafficPointElementService {
     return trafficPointElementVersionRepository.existsBySloid(sloid);
   }
 
+  @PreAuthorize("@countryAndBusinessOrganisationBasedUserAdministrationService.hasUserPermissionsToCreateOrEditTrafficPoint(#servicePointVersions, "
+          + "T(ch.sbb.atlas.kafka.model.user.admin.ApplicationType).SEPODI)")
+  public TrafficPointElementVersion checkPermissionRightsAndSave(TrafficPointElementVersion trafficPointElementVersion, List<ServicePointVersion> servicePointVersions) {
+    return save(trafficPointElementVersion);
+  }
+
   public TrafficPointElementVersion save(TrafficPointElementVersion trafficPointElementVersion) {
+    trafficPointElementValidationService.validateServicePointNumberExists(trafficPointElementVersion.getServicePointNumber());
     return trafficPointElementVersionRepository.save(trafficPointElementVersion);
   }
 
@@ -45,10 +58,20 @@ public class TrafficPointElementService {
     trafficPointElementVersionRepository.deleteById(id);
   }
 
-  public void updateTrafficPointElementVersion(TrafficPointElementVersion edited) {
-    List<TrafficPointElementVersion> dbVersions = findBySloidOrderByValidFrom(edited.getSloid());
-    TrafficPointElementVersion current = BasePointUtility.getCurrentPointVersion(dbVersions, edited);
-    List<VersionedObject> versionedObjects = versionableService.versioningObjects(current, edited,
+  @PreAuthorize("@countryAndBusinessOrganisationBasedUserAdministrationService.hasUserPermissionsToCreateOrEditTrafficPoint(#currentVersions, "
+          + "T(ch.sbb.atlas.kafka.model.user.admin.ApplicationType).SEPODI)")
+  public void checkPermissionRightsAndUpdate(TrafficPointElementVersion currentVersionTPEV, TrafficPointElementVersion editedVersionTPEV, List<ServicePointVersion> currentVersions) {
+    updateTrafficPointElementVersion(currentVersionTPEV, editedVersionTPEV);
+  }
+
+  public void updateTrafficPointElementVersion(TrafficPointElementVersion currentVersion, TrafficPointElementVersion editedVersion) {
+    trafficPointElementVersionRepository.incrementVersion(currentVersion.getSloid());
+    if (editedVersion.getVersion() != null && !currentVersion.getVersion().equals(editedVersion.getVersion())) {
+      throw new StaleObjectStateException(ServicePointVersion.class.getSimpleName(), "version");
+    }
+
+    List<TrafficPointElementVersion> dbVersions = findBySloidOrderByValidFrom(currentVersion.getSloid());
+    List<VersionedObject> versionedObjects = versionableService.versioningObjects(currentVersion, editedVersion,
         dbVersions);
     versionableService.applyVersioning(TrafficPointElementVersion.class, versionedObjects, this::save, this::deleteById);
   }
